epochStart: split rewardsCreator into baseRewardsCreator and rewardsCreator, base to be reused for v2
